entities corregidas y services completados

+++ src/estudiante/estudiante.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { EstudianteEntity } from './estudiante.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Long, Repository } from 'typeorm';
-import { EstudianteDTO } from './estudiante.DTO';
+import { EstudianteDTO } from './estudiante.dto';
 import { ExceptionsHandler } from '@nestjs/core/exceptions/exceptions-handler';
 @Injectable()
 export class EstudianteService {
@@ -15,9 +15,8 @@ export class EstudianteService {
 
     
   async crearEstudiante(estudianteDto: EstudianteDTO ): Promise<EstudianteEntity> {
-    if (estudianteDto.promedio>3.2 && estudianteDto.promedio>=4) {
+    if (estudianteDto.promedio>3.2 && estudianteDto.semestre>=4) {
         const estudiante = this.estudianteRepository.create({
-            id:estudianteDto.id,
             cedula:estudianteDto.cedula,
             semestre:estudianteDto.semestre,
             programa:estudianteDto.programa,
@@ -34,11 +33,15 @@ export class EstudianteService {
   }
   async eliminarEstudiante(id: Long): Promise<void> {
     const estudiante = await this.estudianteRepository.findOne({
-      where: { id }
+      where: { id: Number(id)   }
     });
     if (!estudiante) {
       throw new Error('Estudiante no encontrado');
     }
+    if (estudiante.proyectos.length > 0) {
+      throw new Error('El estudiante aun tiene proyectos activos');
+    }
+  
     await this.estudianteRepository.remove(estudiante);
   }
